+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-oauth-application-headless-server/src/main/java/com/liferay/jethr0/dalo/BuildParameterDALO.java
@@ -14,9 +14,9 @@
 
 package com.liferay.jethr0.dalo;
 
-import com.liferay.jethr0.builds.Build;
-import com.liferay.jethr0.builds.parameter.BuildParameter;
-import com.liferay.jethr0.builds.parameter.BuildParameterFactory;
+import com.liferay.jethr0.build.Build;
+import com.liferay.jethr0.build.parameter.BuildParameter;
+import com.liferay.jethr0.build.parameter.BuildParameterFactory;
 
 import java.util.ArrayList;
 import java.util.List;
